requirements.txt
pyrogram==2.0.106
openai==1.54.0
supabase==2.3.4
python-dotenv==1.0.1
